+++ apps/website/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from apps.website.views import HomePage, AboutPage, ProjectsPage, ProjectDetailPage
+from apps.website.views import HomePage, AboutPage, ProjectsPage, ProjectDetailPage, ContactPage
 
 app_name = 'website'
 
@@ -9,4 +9,5 @@ urlpatterns = [
     path('about/', AboutPage.as_view(), name='about'),
     path('projects/', ProjectsPage.as_view(), name='projects'),
     path('projects/<str:project_slug>', ProjectDetailPage.as_view(), name='project_detail'),
+    path('contact/', ContactPage.as_view(), name='contact'),
 ]